Partial implementation of two line animation
I added a new module called F1D_Enrichment_Calc.py. so far this module
consists of an animation function similar to the one found in
F1d_Animation.py, except that it plots another line with a different
viscosity. The code runs, but the second line (red) has a pathology at
the right edge of the graph. I can’t understand why one of the lines
has it and the other doesn’t.