Alex edits
Alex used a slightly different method to count the pixels and fit an
equation to the resulting data. The next step will require replacing
the dummy function with the kinetics equation from Watson and Price,
and rescaling the data and function so that the results have correct
units. Almost done!